use kyc url retrieved from stack information
