+++ pkg/environment/environment.go
@@ -50,6 +50,7 @@ type Environment struct {
 	RelayURL      []string
 	ActivationURL []string
 	GraphQL       []string
+	KycURL        string
 
 	// private vlan to join
 	// if set, zos will use this as its priv vlan
@@ -124,6 +125,7 @@ var (
 			"https://graphql.dev.grid.tf/graphql",
 			"https://graphql.02.dev.grid.tf/graphql",
 		},
+		KycURL: "https://kyc1.gent01.dev.grid.tf",
 	}
 
 	envTest = Environment{
@@ -146,6 +148,7 @@ var (
 			"https://graphql.test.grid.tf/graphql",
 			"https://graphql.02.test.grid.tf/graphql",
 		},
+		KycURL: "",
 	}
 
 	envQA = Environment{
@@ -168,6 +171,7 @@ var (
 			"https://graphql.qa.grid.tf/graphql",
 			"https://graphql.02.qa.grid.tf/graphql",
 		},
+		KycURL: "",
 	}
 
 	envProd = Environment{
@@ -193,6 +197,7 @@ var (
 			"https://graphql.grid.tf/graphql",
 			"https://graphql.02.grid.tf/graphql",
 		},
+		KycURL: "",
 	}
 )
